Move model to internal/pkg/service/buffer/model Fix lint
